requirements.txt
PySimpleGUI
requests
Pillow
numpy
opencv-python
imutils

